refactor(gateway): move routing config from java beans to yml file
