+++ backend/mazad-gateway/src/main/java/com/mazad/mazadgateway/filters/ApiKeyFilter.java
@@ -5,7 +5,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.gateway.filter.GatewayFilter;
 import org.springframework.cloud.gateway.filter.GatewayFilterChain;
+import org.springframework.cloud.gateway.filter.GlobalFilter;
+import org.springframework.core.Ordered;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.server.RequestPath;
 import org.springframework.http.server.reactive.ServerHttpResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
@@ -16,20 +19,37 @@ import java.util.List;
 
 
 @Component
-public class ApiKeyFilter implements GatewayFilter {
+public class ApiKeyFilter implements GlobalFilter, Ordered {
 
+    private static final Logger log = LoggerFactory.getLogger(ApiKeyFilter.class);
     @Value("${API_KEY}")
     private String API_KEY;
 
     @Override
     public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
         String apiKey = exchange.getRequest().getHeaders().getFirst("X-API-KEY");
+        if (!checkProtectedPaths(exchange.getRequest().getURI().getPath()))
+            return chain.filter(exchange);
+
         if (apiKey == null || !apiKey.equals(API_KEY)) {
-            System.out.println("Invalid api key: " + apiKey);
+            log.warn("[ApiKeyFilter] API KEY: blocked");
             ServerHttpResponse response = exchange.getResponse();
             response.setStatusCode(HttpStatus.UNAUTHORIZED);
             return response.setComplete();
         }
+        log.info("[ApiKeyFilter] API KEY: allowed");
         return chain.filter(exchange);
+    }
+// /api/v1/items/**
+// /api/v1/categories/**
+
+    private boolean checkProtectedPaths(String path) {
+
+        return path.startsWith("/api/v1/items") || path.startsWith("/api/v1/categories");
+    }
+
+    @Override
+    public int getOrder() {
+        return 0;
     }
 }
